+++ sklearn/metrics/tests/test_metrics.py
@@ -7,6 +7,11 @@ import numpy as np
 from sklearn import datasets
 from sklearn import svm
 
+from sklearn.preprocessing import LabelBinarizer
+from sklearn.datasets import make_multilabel_classification
+from sklearn.utils import (check_random_state,
+                           shuffle)
+from sklearn.utils.multiclass import unique_labels
 from sklearn.utils.testing import (assert_true,
                                    assert_raises,
                                    assert_equal,
@@ -24,6 +29,8 @@ from sklearn.metrics import (accuracy_score,
                              confusion_matrix,
                              explained_variance_score,
                              f1_score,
+                             fbeta_score,
+                             hamming_loss,
                              hinge_loss,
                              matthews_corrcoef,
                              mean_squared_error,
@@ -498,6 +505,7 @@ def test_losses():
     """Test loss functions"""
     y_true, y_pred, _ = make_prediction(binary=True)
     n_samples = y_true.shape[0]
+    n_classes = np.size(unique_labels(y_true))
 
     # Classification
     # --------------
@@ -512,6 +520,9 @@ def test_losses():
     assert_equal(zero_one_loss(y_true, y_pred, normalize=False), 13)
     assert_almost_equal(zero_one_loss(y_true, y_true), 0.0, 2)
     assert_almost_equal(zero_one_loss(y_true, y_true, normalize=False), 0, 2)
+
+    assert_almost_equal(hamming_loss(y_true, y_pred),
+                        2 * 13. / (n_samples * n_classes), 2)
 
     assert_equal(accuracy_score(y_true, y_pred),
                  1 - zero_one_loss(y_true, y_pred))
@@ -561,41 +572,69 @@ def test_symmetry():
     """Test the symmetry of score and loss functions"""
     y_true, y_pred, _ = make_prediction(binary=True)
 
-    # symmetric
-    assert_equal(accuracy_score(y_true, y_pred),
-                 accuracy_score(y_pred, y_true))
+    # Symmetric metric
+    for metric in [accuracy_score,
+                   zero_one_loss,
+                   lambda y1, y2: zero_one_loss(y1, y2, normalize=False),
+                   hamming_loss,
+                   f1_score,
+                   matthews_corrcoef,
+                   mean_squared_error,
+                   mean_absolute_error]:
 
+        assert_equal(metric(y_true, y_pred),
+                     metric(y_pred, y_true),
+                     msg="%s is not symetric" % metric)
+
+    # Not symmetric metrics
+    for metric in [precision_score,
+                   recall_score,
+                   lambda y1, y2: fbeta_score(y1, y2, beta=0.5),
+                   lambda y1, y2: fbeta_score(y1, y2, beta=2),
+                   explained_variance_score,
+                   r2_score]:
+
+        assert_true(metric(y_true, y_pred) != metric(y_pred, y_true),
+                    msg="%s seems to be symetric" % metric)
+
+    # Deprecated metrics
     with warnings.catch_warnings(True):
         # Throw deprecated warning
         assert_equal(zero_one(y_true, y_pred),
                      zero_one(y_pred, y_true))
 
-        assert_almost_equal(zero_one(y_true, y_pred, normalize=False),
-                            zero_one(y_pred, y_true, normalize=False), 2)
+        assert_equal(zero_one(y_true, y_pred, normalize=False),
+                     zero_one(y_pred, y_true, normalize=False))
 
-    assert_equal(zero_one_loss(y_true, y_pred),
-                 zero_one_loss(y_pred, y_true))
-
-    assert_equal(zero_one_loss(y_true, y_pred, normalize=False),
-                 zero_one_loss(y_pred, y_true, normalize=False))
-
-    with warnings.catch_warnings(True):
-    # Throw deprecated warning
         assert_equal(zero_one_score(y_true, y_pred),
                      zero_one_score(y_pred, y_true))
 
-    assert_almost_equal(mean_squared_error(y_true, y_pred),
-                        mean_squared_error(y_pred, y_true))
 
-    assert_almost_equal(mean_absolute_error(y_true, y_pred),
-                        mean_absolute_error(y_pred, y_true))
+def test_sample_order_invariance():
+    y_true, y_pred, _ = make_prediction(binary=True)
 
-    # not symmetric
-    assert_true(explained_variance_score(y_true, y_pred) !=
-                explained_variance_score(y_pred, y_true))
-    assert_true(r2_score(y_true, y_pred) !=
-                r2_score(y_pred, y_true))
-    # FIXME: precision and recall aren't symmetric either
+    y_true_shuffle, y_pred_shuffle = shuffle(y_true, y_pred,
+                                             random_state=0)
+
+    for metric in [accuracy_score,
+                   hamming_loss,
+                   zero_one_loss,
+                   lambda y1, y2: zero_one_loss(y1, y2, normalize=False),
+                   precision_score,
+                   recall_score,
+                   f1_score,
+                   lambda y1, y2: fbeta_score(y1, y2, beta=2),
+                   lambda y1, y2: fbeta_score(y1, y2, beta=0.5),
+                   matthews_corrcoef,
+                   mean_absolute_error,
+                   mean_squared_error,
+                   explained_variance_score,
+                   r2_score]:
+
+        assert_almost_equal(metric(y_true, y_pred),
+                            metric(y_true_shuffle, y_pred_shuffle),
+                            err_msg="%s is not sample order invariant"
+                                    % metric)
 
 
 def test_hinge_loss_binary():
@@ -682,3 +721,190 @@ def test_multioutput_regression_invariance_to_dimension_shuffling():
             perm = np.random.permutation(n_dims)
             assert_almost_equal(error,
                                 metric(y_true[:, perm], y_pred[:, perm]))
+
+
+def test_multilabel_representation_invariance():
+
+    MULTILABELS_METRICS = [hamming_loss,
+                           zero_one_loss,
+                           accuracy_score]
+
+    # Generate some data
+    n_classes = 4
+    n_samples = 50
+    _, y1 = make_multilabel_classification(n_features=1, n_classes=n_classes,
+                                           random_state=0, n_samples=n_samples)
+    _, y2 = make_multilabel_classification(n_features=1, n_classes=n_classes,
+                                           random_state=1, n_samples=n_samples)
+
+    # Be sure to have at least one empty label
+    y1 += ([], )
+    y2 += ([], )
+
+    # NOTE: The "sorted" trick is necessary to shuffle labels, because it
+    # allows to return the shuffled tuple.
+    py_random_state = random.Random(0)
+    shuffled = lambda x: sorted(x, key=lambda *args: py_random_state.random())
+    y1_shuffle = [shuffled(x) for x in y1]
+    y2_shuffle = [shuffled(x) for x in y2]
+
+    # Let's have redundant label
+    y1_redundant = [x * py_random_state.randint(1, 3) for x in y1]
+    y2_redundant = [x * py_random_state.randint(1, 3) for x in y2]
+
+    # Binary indicator matrix format
+    lb = LabelBinarizer().fit([range(n_classes)])
+    y1_binary_indicator = lb.transform(y1)
+    y2_binary_indicator = lb.transform(y2)
+
+    y1_shuffle_binary_indicator = lb.transform(y1_shuffle)
+    y2_shuffle_binary_indicator = lb.transform(y2_shuffle)
+
+    for metric in MULTILABELS_METRICS:
+        measure = metric(y1, y2)
+
+        # Check representation invariance
+        assert_almost_equal(measure,
+                            metric(y1_binary_indicator, y2_binary_indicator),
+                            err_msg="%s failed representation invariance  "
+                                    "between list of list of labels format "
+                                    "and dense binary indicator format."
+                                    % metric)
+
+        # Check invariance with redundant labels with list of labels
+        assert_almost_equal(measure,
+                            metric(y1, y2_redundant),
+                            err_msg="%s failed rendundant label invariance"
+                                    % metric)
+
+        assert_almost_equal(measure,
+                            metric(y1_redundant, y2_redundant),
+                            err_msg="%s failed rendundant label invariance"
+                                    % metric)
+
+        assert_almost_equal(measure,
+                            metric(y1_redundant, y2),
+                            err_msg="%s failed rendundant label invariance"
+                                    % metric)
+
+        # Check shuffling invariance with list of labels
+        assert_almost_equal(measure,
+                            metric(y1_shuffle, y2_shuffle),
+                            err_msg="%s failed shuffling invariance "
+                                    "with list of list of labels format."
+                                    % metric)
+
+        # Check shuffling invariance with dense binary indicator matrix
+        assert_almost_equal(measure,
+                            metric(y1_shuffle_binary_indicator,
+                                   y2_shuffle_binary_indicator),
+                            err_msg="%s failed shuffling invariance "
+                                    " with dense binary indicator format."
+                                    % metric)
+
+        # Check invariance with mix input representation
+        assert_almost_equal(measure,
+                            metric(y1,
+                                   y2_binary_indicator),
+                            err_msg="%s failed mix input representation"
+                                    "invariance: y_true in list of list of "
+                                    "labels format and y_pred in dense binary"
+                                    "indicator format"
+                                    % metric)
+
+        assert_almost_equal(measure,
+                            metric(y1_binary_indicator,
+                                   y2),
+                            err_msg="%s failed mix input representation"
+                                    "invariance: y_true in dense binary "
+                                    "indicator format and y_pred in list of "
+                                    "list of labels format."
+                                    % metric)
+
+
+def test_multilabel_zero_one_loss():
+    # Dense label indicator matrix format
+    y1 = np.array([[0.0, 1.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+    y2 = np.array([[0.0, 0.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+
+    assert_equal(0.5, zero_one_loss(y1, y2))
+    assert_equal(0.0, zero_one_loss(y1, y1))
+    assert_equal(0.0, zero_one_loss(y2, y2))
+    assert_equal(1.0, zero_one_loss(y2, np.logical_not(y2)))
+    assert_equal(1.0, zero_one_loss(y1, np.logical_not(y1)))
+    assert_equal(1.0, zero_one_loss(y1, np.zeros(y1.shape)))
+    assert_equal(1.0, zero_one_loss(y2, np.zeros(y1.shape)))
+
+    # List of tuple of label
+    y1 = [(1, 2,),
+          (0, 2,)]
+
+    y2 = [(2,),
+          (0, 2,)]
+
+    assert_equal(0.5, zero_one_loss(y1, y2))
+    assert_equal(0.0, zero_one_loss(y1, y1))
+    assert_equal(0.0, zero_one_loss(y2, y2))
+    assert_equal(1.0, zero_one_loss(y2, [(), ()]))
+    assert_equal(1.0, zero_one_loss(y2, [tuple(), (10, )]))
+
+
+def test_multilabel_hamming_loss():
+    # Dense label indicator matrix format
+    y1 = np.array([[0.0, 1.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+    y2 = np.array([[0.0, 0.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+
+    assert_equal(1 / 6., hamming_loss(y1, y2))
+    assert_equal(0.0, hamming_loss(y1, y1))
+    assert_equal(0.0, hamming_loss(y2, y2))
+    assert_equal(1.0, hamming_loss(y2, np.logical_not(y2)))
+    assert_equal(1.0, hamming_loss(y1, np.logical_not(y1)))
+    assert_equal(4. / 6, hamming_loss(y1, np.zeros(y1.shape)))
+    assert_equal(0.5, hamming_loss(y2, np.zeros(y1.shape)))
+
+    # List of tuple of label
+    y1 = [(1, 2,),
+          (0, 2,)]
+
+    y2 = [(2,),
+          (0, 2,)]
+
+    assert_equal(1 / 6., hamming_loss(y1, y2))
+    assert_equal(0.0, hamming_loss(y1, y1))
+    assert_equal(0.0, hamming_loss(y2, y2))
+    assert_equal(0.75, hamming_loss(y2, [(), ()]))
+    assert_equal(0.625, hamming_loss(y1, [tuple(), (10, )]))
+    assert_almost_equal(0.1818, hamming_loss(y2, [tuple(), (10, )],
+                                             classes=np.arange(11)), 2)
+
+
+def test_multilabel_accuracy_score():
+    # Dense label indicator matrix format
+    y1 = np.array([[0.0, 1.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+    y2 = np.array([[0.0, 0.0, 1.0],
+                   [1.0, 0.0, 1.0]])
+
+    assert_equal(0.5, accuracy_score(y1, y2))
+    assert_equal(1.0, accuracy_score(y1, y1))
+    assert_equal(1.0, accuracy_score(y2, y2))
+    assert_equal(0.0, accuracy_score(y2, np.logical_not(y2)))
+    assert_equal(0.0, accuracy_score(y1, np.logical_not(y1)))
+    assert_equal(0.0, accuracy_score(y1, np.zeros(y1.shape)))
+    assert_equal(0.0, accuracy_score(y2, np.zeros(y1.shape)))
+
+    # List of tuple of label
+    y1 = [(1, 2,),
+          (0, 2,)]
+
+    y2 = [(2,),
+          (0, 2,)]
+
+    assert_equal(0.5, accuracy_score(y1, y2))
+    assert_equal(1.0, accuracy_score(y1, y1))
+    assert_equal(1.0, accuracy_score(y2, y2))
+    assert_equal(0.0, accuracy_score(y2, [(), ()]))