+++ doc/whats_new.rst
@@ -88,6 +88,17 @@ List of contributors for release 0.13.1 by number of commits.
  *  1  `Vlad Niculae`_
  *  1  `Yaroslav Halchenko`_
 
+=======
+   - :func:`metrics.accuracy_score` and :func:`metrics.zero_one_loss` support
+     multi-label classification.  A new metric :func:`metrics.hamming_loss` is
+     added with mullti-label support.
+>>>>>>> Update what's new
+=======
+   - :func:`metrics.accuracy_score`, :func:`metrics.zero_one_loss` support
+     multi-label classification and a new metric :func:`metrics.hamming_loss`
+     is added with multi-label support by `Arnaud Joly`_.
+
+>>>>>>> What's new : add credit
 
 .. _changes_0_13: